requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
cffi==1.11.5
cryptography==2.3.1
future==0.16.0
idna==2.7
pycparser==2.19
PySocks==1.6.8
python-telegram-bot==11.1.0
selenium==3.14.1
six==1.11.0
urllib3==1.23
#https://www.joecolantonio.com/how-to-install-phantomjs/

